Time: 84 ms (21.51%), Space: 162.8 MB (31.03%) - LeetHub

+++ 3867-sum-of-gcd-of-formed-pairs/3867-sum-of-gcd-of-formed-pairs.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long gcdSum(vector<int>& nums) {
-        vector<long long>prefixGcd;
+        vector<int>prefixGcd;
         int n = nums.size();
 
         int maxEle = 0;
@@ -9,11 +9,14 @@ public:
             maxEle = max(maxEle,nums[i]);
 
             prefixGcd.push_back(gcd(nums[i],maxEle));
+            // instead you can also update the value in the same place 
+            // nums[i] = gcd(maxEle,nums[i]);
+            // it will reduce the s(0) = o(1) instead of o(n)
         }
 
         sort(begin(prefixGcd),end(prefixGcd));
-        long long s = 0;
-        long long e = prefixGcd.size()-1;
+        int s = 0;
+        int e = prefixGcd.size()-1;
         long long res = 0;
         while(s<e){
             long long curGcd = gcd(prefixGcd[s],prefixGcd[e]);
